scripts/train.py
"""
Training Script
Main entry point for training the PPO agent.

Usage: python scripts/train.py
"""

import torch
import yaml
import sys
sys.path.append('src')

from models.actor_critic import ActorCritic
from algorithms.ppo import PPO
from algorithms.replay_buffer import RolloutBuffer
from environment.spider_env import SpiderBotEnv
from utils.logger import Logger


def train():
    """
    Main training loop.
    
    TODO: Implement training loop
    1. Load configs
    2. Create environment
    3. Create actor-critic network
    4. Create PPO agent
    5. Training loop:
        - Collect rollouts (n_steps)
        - Update policy
        - Log metrics
        - Save checkpoints
    
    Pseudocode:
    for step in range(total_timesteps):
        action, log_prob, value = agent.get_action(obs)
        next_obs, reward, done = env.step(action)
        buffer.add(obs, action, reward, value, log_prob, done)
        
        if buffer.is_full():
            metrics = agent.update(buffer)
            buffer.reset()
    """
    
    # TODO: Load configurations
    # with open('config/training_config.yaml') as f:
    #     config = yaml.safe_load(f)
    
    # TODO: Initialize environment, model, agent
    
    # TODO: Training loop
    
    print("Training complete!")


if __name__ == "__main__":
    train()
